+++ packages/@livestore/adapter-web/src/web-worker/client-session/persisted-adapter.ts
@@ -4,12 +4,13 @@ import {
   liveStoreVersion,
   makeClientSession,
   StoreInterrupted,
+  sessionChangesetMetaTable,
   UnexpectedError,
 } from '@livestore/common'
 // TODO bring back - this currently doesn't work due to https://github.com/vitejs/vite/issues/8427
 // NOTE We're using a non-relative import here for Vite to properly resolve the import during app builds
 // import LiveStoreSharedWorker from '@livestore/adapter-web/internal-shared-worker?sharedworker'
-import { EventSequenceNumber, SystemTables } from '@livestore/common/schema'
+import { EventSequenceNumber } from '@livestore/common/schema'
 import { sqliteDbFactory } from '@livestore/sqlite-wasm/browser'
 import { loadSqlite3Wasm } from '@livestore/sqlite-wasm/load-wasm'
 import { isDevEnv, shouldNeverHappen, tryAsFunctionAndNew } from '@livestore/utils'
@@ -67,7 +68,9 @@ export type WebAdapterOptions = {
    */
   sharedWorker:
     | ((options: { name: string }) => globalThis.SharedWorker)
-    | (new (options: { name: string }) => globalThis.SharedWorker)
+    | (new (options: {
+        name: string
+      }) => globalThis.SharedWorker)
   /**
    * Specifies where to persist data for this adapter
    */
@@ -366,17 +369,15 @@ export const makePersistedAdapter =
 
       // We're restoring the leader head from the SESSION_CHANGESET_META_TABLE, not from the eventlog db/table
       // in order to avoid exporting/transferring the eventlog db/table, which is important to speed up the fast path.
-      const initialLeaderHeadRes = sqliteDb.select<{
-        seqNumGlobal: EventSequenceNumber.GlobalEventSequenceNumber
-        seqNumClient: EventSequenceNumber.ClientEventSequenceNumber
-      }>(
-        `select seqNumGlobal, seqNumClient from ${SystemTables.SESSION_CHANGESET_META_TABLE} order by seqNumGlobal desc, seqNumClient desc limit 1`,
-      )[0]
+      const initialLeaderHeadRes = sqliteDb.select(
+        sessionChangesetMetaTable.select('seqNumClient', 'seqNumGlobal', 'seqNumRebaseGeneration').first(),
+      )
 
       const initialLeaderHead = initialLeaderHeadRes
         ? EventSequenceNumber.make({
             global: initialLeaderHeadRes.seqNumGlobal,
             client: initialLeaderHeadRes.seqNumClient,
+            rebaseGeneration: initialLeaderHeadRes.seqNumRebaseGeneration,
           })
         : EventSequenceNumber.ROOT