+++ packages/@livestore/common/src/leader-thread/leader-worker-devtools.ts
@@ -48,9 +48,8 @@ export const bootDevtools = (options: DevtoolsOptions) =>
               )
 
           const syncState = yield* syncProcessor.syncState
-          const mergeCounter = syncProcessor.getMergeCounter()
 
-          yield* syncProcessor.pull({ cursor: { mergeCounter, eventNum: syncState.localHead } }).pipe(
+          yield* syncProcessor.pull({ cursor: syncState.localHead }).pipe(
             Stream.tap(({ payload }) => sendMessage(Devtools.Leader.SyncPull.make({ payload, liveStoreVersion }))),
             Stream.runDrain,
             Effect.forkScoped,