+++ tests/integration/src/tests/playwright/fixtures/devtools/two-stores/Root.tsx
@@ -1,7 +1,6 @@
 import { makePersistedAdapter } from '@livestore/adapter-web'
 import LiveStoreSharedWorker from '@livestore/adapter-web/shared-worker?sharedworker'
-import { LiveStoreProvider, useStore } from '@livestore/react'
-import React from 'react'
+import { LiveStoreProvider } from '@livestore/react'
 import { unstable_batchedUpdates as batchedUpdates } from 'react-dom'
 
 import LiveStoreWorkerNotes from './livestore-notes.worker.ts?worker'
@@ -10,8 +9,6 @@ import { schema as schemaNotes } from './schema-notes.js'
 import { schema as schemaTodos } from './schema-todos.js'
 
 const Notes = () => {
-  const { store } = useStore()
-
   return (
     <div>
       <h1>Notes</h1>
@@ -20,8 +17,6 @@ const Notes = () => {
 }
 
 const Todos = () => {
-  const { store } = useStore()
-
   return (
     <div>
       <h1>Todos</h1>